sql_validate_graduation will generate the standard sql query for pulling Graduation Validations from Edify. Validations are done through Edify.

sql_validate_graduation will generate the standard sql query for pulling Graduation Validations from Edify. Validations are done through Edify.

Usage

sql_validate_graduation(data_source, name, context)

Arguments

data_source

The server you will be full from. Currently, Edify.

name

The name you want the SQL file to have in your sql folder. This is a string and must be provided in quotes (e.g., "your_filename.sql").

context

The context of your project. "project", "shiny", and "sandbox". The default is "project".

Value

A sql file in your SQL folder